% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HP_Thinning.R
\name{HP.thinnig}
\alias{HP.thinnig}
\title{Hawkes Process Simulation with Thinning Algorithm}
\usage{
HP.thinnig(T, lam0, phi)
}
\arguments{
\item{lam0}{The base intensity function of Hawkes Process}

\item{phi}{The excitation function of Hawkes Process}
}
\description{
This function simulates the event times of Hawkes Process(HP).

It uses the thinning algorigthm (Lewis and Shedler 1979)
}
\examples{
HP.thinnig(T=20,lam0 = function(x) {0*x+1},phi = function(x) {3.9*exp(-4*x)})
HP.thinnig(T=20,lam0 = function(x) {0*x+1},phi = function(x) {2000*(x+87)^(-2.33)})

}
