% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HP_Composition_R.R
\name{HP_Composition_R}
\alias{HP_Composition_R}
\title{Hawkes Process Simulation with Composition Algorithm with Runuran}
\usage{
HP_Composition_R(n, T, lam0, phi)
}
\arguments{
\item{n}{The number of independent Hawkes Process simulations}

\item{lam0}{The base intensity function of Hawkes Process}

\item{phi}{The excitation function of Hawkes Process}
}
\description{
This function simulates the event times of Hawkes Process(HP).

It is based on cluster-based representation of Hawkes Process

The immigrant process and the offspring processes of HP are

independent Non-Homogeneous Poisson Processes (NHPP).

The Hawkes Process is the combination of these NHPPs with

different parent event times.

Therefore it is a combination algorithm.

The NHPPs are are simulated using transformation method with order statistics

The R package RUNURAN is used for the numerical inversion of

the cumulative intensity function
}
\examples{
HP_Composition_R(n=1,T=20,lam0 = function(x) {0*x+1},phi = function(x) {3.9*exp(-4*x)})
HP_Composition_R(n=10,T=20,lam0 = function(x) {0*x+1},phi = function(x) {2000*(x+87)^(-2.33)})

}
