% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPP_Inv_Order_Stat.R
\name{NHPP.inversion.order.statistics}
\alias{NHPP.inversion.order.statistics}
\title{Non-Homogeneous Possion Process Simulation with Inversion (Transformation) combined with Order Statistics}
\usage{
NHPP.inversion.order.statistics(T, lam)
}
\arguments{
\item{lam}{The intensity function of NHPP}
}
\description{
This function simulates the event times of Non-Homogeneous Possion Process (NHHP).

It uses Inversion (Transformation) Algorithm

It is combined with Order Statistics to prevent the events to be rejected
}
\examples{
NHPP.inversion.order.statistics(T=20,lam = function(x) {0*x+1})
NHPP.inversion.order.statistics(T=20,lam = function(x) {3.9*exp(-4*x)})
NHPP.inversion.order.statistics(T=20,lam = function(x) {2000*(x+87)^(-2.33)})

}
