% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPP_Thinning.R
\name{NHPP.thinnig}
\alias{NHPP.thinnig}
\title{Non-Homogeneous Possion Process Simulation with Thinning Algorithm}
\usage{
NHPP.thinnig(T, lam)
}
\arguments{
\item{lam}{The intensity function of NHPP}
}
\description{
This function simulates the event times of Non-Homogeneous Possion Process (NHHP).

It uses Thinning Algorithm (Lewis and Shedler, 1979)
}
\examples{
NHPP.thinnig(T=20,lam = function(x) {0*x+1})
NHPP.thinnig(T=20,lam = function(x) {2000*(x+87)^(-2.33)})
NHPP.thinnig(T=20,lam = function(x) {3.9*exp(-4*x)})

}
